fix: unify prepare concealed and plaintext reqs to fix mixed scenarios
If a key is set but the request had no `%allot` fields then the request wasn't been prepared for each node. This commit unifies the `prepareConcealedRequest` and `preparePlaintextRequest` — and now walks the request body to identify if there are `%allot`s to process.

This isn't an elegant solution, but given that body sizes are limited it is sufficient until a more thoughtful approach can be implemented.